move bannerLinks to atoms, update imports
