+++ apps/nextjs-website/src/_contents/makeDocs.ts
@@ -4,7 +4,7 @@ import * as RA from 'fp-ts/lib/ReadonlyArray';
 import { docsAssetsPath, docsPath } from '@/config';
 import { Product } from '@/lib/types/product';
 import { parseDoc } from 'gitbook-docs/parseDoc';
-import { BannerLinkProps } from '@/editorialComponents/BannerLink';
+import { BannerLinkProps } from '@/components/atoms/BannerLink/BannerLink';
 
 export type TutorialsDefinition = {
   readonly product: Product;
