--- apps/nextjs-website/src/editorialComponents/BannerLink/index.ts
@@ -1,1 +1,0 @@
-export * from './BannerLink';
